feat(sticker): shared selection gear in both the emoji and sticker selectors
Extract the selection-preferences panel into a shared CustomEmojiSelectionSettings component — it governs the per-chat customEmojiSelection prefs that drive BOTH emoji and sticker advertising — and surface its gear in the sticker selector too, relabeled to cover emojis & stickers. CustomEmojiTab now renders the shared component instead of its inline panel, so there's a single source of truth reachable from either selector.
