+++ packages/client/src/components/chat/CustomEmojiTab.tsx
@@ -4,7 +4,7 @@
 // (in edit mode) renames/deletes them. Rendered into EmojiPicker via its
 // optional `customTab` slot so the picker itself stays generic.
 // ──────────────────────────────────────────────
-import { useCallback, useEffect, useRef, useState, type ChangeEvent } from "react";
+import { useCallback, useRef, useState, type ChangeEvent } from "react";
 import { ImagePlus, Settings, Trash2 } from "lucide-react";
 import {
   useCustomEmojis,
@@ -14,17 +14,8 @@ import {
   useImportCustomEmojis,
 } from "../../hooks/use-custom-emojis";
 import { useConversationCustomEmojis, type ConversationCustomEmoji } from "../../hooks/use-conversation-custom-emojis";
-import { useChat, useUpdateChatMetadata } from "../../hooks/use-chats";
-import { useConnections } from "../../hooks/use-connections";
-import { useChatStore } from "../../stores/chat.store";
-import { parseChatMetadata } from "../../lib/chat-display";
+import { CustomEmojiSelectionSettings } from "./CustomEmojiSelectionSettings";
 import { readImageDimensions, validateDimensionsForKind, slugifyCustomName } from "../../lib/custom-emoji";
-import {
-  normalizeCustomEmojiSelection,
-  CUSTOM_EMOJI_SELECTION_MIN_COUNT,
-  CUSTOM_EMOJI_SELECTION_MAX_COUNT,
-  type CustomEmojiSelectionPrefs,
-} from "@marinara-engine/shared";
 import { showPromptDialog, showConfirmDialog } from "../../lib/app-dialogs";
 import { downloadJsonFile } from "../../lib/download-json";
 import { api } from "../../lib/api-client";
@@ -37,27 +28,11 @@ export function CustomEmojiTab({ onInsert, query }: { onInsert: (token: string) 
   const remove = useDeleteCustomEmoji();
   const importEmojis = useImportCustomEmojis();
   const { list: conversationEmojis } = useConversationCustomEmojis();
-  const activeChatId = useChatStore((s) => s.activeChatId);
-  const { data: activeChat } = useChat(activeChatId);
-  const updateMeta = useUpdateChatMetadata();
-  const { data: connections } = useConnections();
   const fileRef = useRef<HTMLInputElement>(null);
   const importFileRef = useRef<HTMLInputElement>(null);
   const [editing, setEditing] = useState(false);
   const [showSettings, setShowSettings] = useState(false);
   const [error, setError] = useState<string | null>(null);
-
-  const selectionPrefs = normalizeCustomEmojiSelection(parseChatMetadata(activeChat?.metadata).customEmojiSelection);
-  const [maxCountDraft, setMaxCountDraft] = useState(selectionPrefs.maxCount);
-  useEffect(() => setMaxCountDraft(selectionPrefs.maxCount), [selectionPrefs.maxCount]);
-
-  const saveSelectionPrefs = useCallback(
-    (patch: Partial<CustomEmojiSelectionPrefs>) => {
-      if (!activeChatId) return;
-      updateMeta.mutate({ id: activeChatId, customEmojiSelection: { ...selectionPrefs, ...patch } });
-    },
-    [activeChatId, selectionPrefs, updateMeta],
-  );
 
   const list = emojis ?? [];
 
@@ -251,70 +226,7 @@ export function CustomEmojiTab({ onInsert, query }: { onInsert: (token: string) 
         </div>
       </div>
 
-      {showSettings && (
-        <div className="mb-2 rounded-md bg-foreground/5 p-2 ring-1 ring-foreground/10">
-          <p className="mb-1.5 text-[0.6875rem] text-foreground/55">
-            When a character has more custom emojis than the max, how should the ones offered to the model be chosen?
-          </p>
-          <div className="mb-1.5 flex items-center gap-1">
-            {(["semantic", "random", "tool-call"] as const).map((mode) => (
-              <button
-                key={mode}
-                type="button"
-                onClick={() => saveSelectionPrefs({ mode })}
-                className={cn(
-                  "flex-1 rounded px-2 py-1 text-xs capitalize transition-colors",
-                  selectionPrefs.mode === mode
-                    ? "bg-[var(--primary)] text-white"
-                    : "bg-foreground/5 text-foreground/60 ring-1 ring-foreground/10 hover:bg-foreground/10",
-                )}
-              >
-                {mode}
-              </button>
-            ))}
-          </div>
-          <p className="mb-2 text-[0.625rem] text-foreground/40">
-            {selectionPrefs.mode === "semantic"
-              ? "Offers the emojis most relevant to the recent conversation (falls back to random if the local embedder is unavailable)."
-              : selectionPrefs.mode === "random"
-                ? "Offers a random set for each reply."
-                : "A model call picks the fitting emojis each reply — choose a capable connection below. Falls back to semantic if it's unset or fails."}
-          </p>
-          {selectionPrefs.mode === "tool-call" && (
-            <div className="mb-2">
-              <select
-                value={selectionPrefs.toolConnectionId ?? ""}
-                onChange={(e) => saveSelectionPrefs({ toolConnectionId: e.target.value || null })}
-                className="w-full rounded bg-foreground/5 px-2 py-1 text-xs text-foreground ring-1 ring-foreground/10 focus:outline-none focus:ring-[var(--primary)]"
-              >
-                <option value="">Select a connection…</option>
-                {((connections ?? []) as Array<{ id: string; name?: string }>).map((connection) => (
-                  <option key={connection.id} value={connection.id}>
-                    {connection.name || connection.id}
-                  </option>
-                ))}
-              </select>
-              {!selectionPrefs.toolConnectionId && (
-                <p className="mt-1 text-[0.625rem] text-amber-400/80">
-                  No connection set — this falls back to semantic selection.
-                </p>
-              )}
-            </div>
-          )}
-          <label className="flex items-center justify-between gap-2 text-xs text-foreground/60">
-            <span>Max emojis offered</span>
-            <input
-              type="number"
-              min={CUSTOM_EMOJI_SELECTION_MIN_COUNT}
-              max={CUSTOM_EMOJI_SELECTION_MAX_COUNT}
-              value={maxCountDraft}
-              onChange={(e) => setMaxCountDraft(Number(e.target.value))}
-              onBlur={() => saveSelectionPrefs({ maxCount: maxCountDraft })}
-              className="w-16 rounded bg-foreground/5 px-2 py-1 text-right text-foreground ring-1 ring-foreground/10 focus:outline-none focus:ring-[var(--primary)]"
-            />
-          </label>
-        </div>
-      )}
+      {showSettings && <CustomEmojiSelectionSettings />}
 
       {error && <p className="mb-2 px-1 text-[0.6875rem] text-red-400">{error}</p>}
